add: made components from the project list and a individual project
